DG_backend/requirements.txt
fastapi 
uvicorn 
motor 
beanie 
passlib[bcrypt] 
python-jose 
pydantic-settings